refaactor(useMessageHandling): split login into useOutsideInputClick
